zmqstat: fix for undefined initial state of a variable
The variable which is supposed to hold the "todo" recipients from the queue file
is not initialized as an array reference.

Therefore, in peculiar cases where the record is missing in the queue file,
zmqstat blows up halfway when trying to dereference as an array something which is undefined.

A case is when a queue object is moved to the hold queue after having
delivered the message to all the recipients but before successful deletion from the queue.

The output then stops at the first problematic queue file with an error
such as
"Can't use an undefined value as an ARRAY reference at /opt/zimbra/libexec/zmqstat line 159"

Initializing it as an empty array reference at the beginning prevents the problem.